moving code into RoutingExtension updating services.yml
