+++ DependencyInjection/Compiler/RouterExtensionCompilerPass.php
@@ -14,9 +14,9 @@ class RouterExtensionCompilerPass implements CompilerPassInterface
         // automatic routing parameters
         $container->setParameter('twig.extension.routing.class', 'Egzakt\\SystemBundle\\Extensions\\RoutingExtension');
         $container->setParameter('router.options.generator_base_class', 'Egzakt\\SystemBundle\\Lib\\RouterUrlGenerator');
-        $container->findDefinition('twig.extension.routing')->addMethodCall('setAutoParametersHandler', array(
+        /*$container->findDefinition('twig.extension.routing')->addMethodCall('setAutoParametersHandler', array(
             new Reference('egzakt_system.router_auto_parameters_handler')
-        ));
+        ));*/
 
         // i18n loader override
         $container->setParameter('jms_i18n_routing.loader.class', 'Egzakt\\SystemBundle\\Routing\\Loader');
@@ -24,5 +24,13 @@ class RouterExtensionCompilerPass implements CompilerPassInterface
         $container->findDefinition('jms_i18n_routing.loader')->addMethodCall('setDatabaseConnection', array(
             new Reference('database_connection')
         ));
+
+        $definition = $container->getDefinition('egzakt_system.entity_route');
+        $taggedServices = $container->findTaggedServiceIds('egzakt_system.entity_route');
+
+        foreach ($taggedServices as $id => $attributes) {
+            $definition->addMethodCall('registerMapping', array(new Reference($id)));
+        }
+
     }
 }
